limpieza parcial

+++ lmf-server-nest/src/main/upload.controller.ts
@@ -3,7 +3,6 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import { InjectRepository } from '@nestjs/typeorm';
 import { diskStorage } from 'multer';
 import { SongEntity } from 'src/_entities/song.entity';
-import { TsFeatureExtractorService } from 'src/_services/ts-feature-extractor.service';
 import { Repository } from 'typeorm';
 import { GzipConverterService } from 'src/_services/gzip-converter.service';
 import { Mp3ValidationPipe } from 'src/_pipes/mp3-validation.pipe';
@@ -14,7 +13,6 @@ import * as path from 'path';
 @Controller('upload')
 export class UploadController {
   constructor(
-    private readonly tsFeatureExtractor: TsFeatureExtractorService,
     private readonly gzipConverter: GzipConverterService,
     @InjectRepository(SongEntity)
     private readonly songRepository: Repository<SongEntity>
@@ -43,13 +41,13 @@ export class UploadController {
       }
     });
 
-    /*if (existingSong) {
+    if (existingSong) {
       const filePath = path.resolve('../uploads', file.filename);
       if (fs.existsSync(filePath))
         await fs.unlinkSync(filePath);
       console.error('Canción duplicada:', file.filename);
       throw new HttpException('Duplicated song', HttpStatus.FORBIDDEN)
-    } else {*/
+    } else {
       const name = body.name;
       const row = await this.songRepository.save(this.songRepository.create({
         name: name,
@@ -60,7 +58,7 @@ export class UploadController {
         message: 'uploaded successful',
         row
       }
-    //}
+    }
   }
 
 }
